FluxDiffUtils Documentation

@includedocs.md

Index

FluxDiffUtils.banded_function_evals

FluxDiffUtils.banded_function_evals!

FluxDiffUtils.hadamard_jacobian

FluxDiffUtils.hadamard_jacobian!

FluxDiffUtils.hadamard_sum

FluxDiffUtils.hadamard_sum!

Functions

FluxDiffUtils.banded_function_evals! — Method

banded_function_evals!(A,mat_fun::Fxn, U, Fargs ...) where Fxn

Mutating version of banded_function_evals.

FluxDiffUtils.banded_function_evals — Method

banded_function_evals(mat_fun::Fxn, U, Fargs...)

Computes block-banded matrix whose bands are entries of matrix-valued function evals (e.g., a Jacobian). Returns SMatrix whose blocks correspond to function components evaluated at values of U.

Example:

julia> mat_fun(U) = [U[1] U[2]; U[2] U[1]]
julia> U = (randn(10),randn(10))
julia> banded_function_evals(mat_fun,U)

FluxDiffUtils.hadamard_jacobian! — Method

hadamard_jacobian!(A::SMatrix{N,N},
                   A_list::NTuple{Nd,AbstractArray},
                   hadamard_product_type::Symbol, dF::Fxn, U,
                   Fargs...; skip_index=(i,j)->false) where {N,Nd,Fxn}

Mutating version of hadamard_jacobian. A = matrix for storing Jacobian output, with each entry storing a block of the Jacobian.

FluxDiffUtils.hadamard_jacobian — Method

hadamard_jacobian(A_list, dF::Fxn, U, Fargs...;
                  skip_index=(i,j)->false) where {N,Fxn}

hadamard_jacobian(A_list, hadamard_product_type, dF::Fxn, U,
                  Fargs...; skip_index=(i,j)->false) where {N,Fxn}

Computes Jacobian of the flux differencing term ∑_i sum(Ai.*Fi). Outputs array whose entries are blocks of the Jacobian matrix corresponding to components of flux vectors.

Inputs:

A_list = tuple of operators

(optional) hadamard_product_type = :skew, :sym. Specifies if Ai.*Fi is skew or symmetric.

dF = Jacobian of the flux function F(uL,uR) with respect to uR

U = solution at which to evaluate the Jacobian

Fargs = extra args for dF(uL,uR)

(optional) skip_index(i,j) = optional function to skip computation of (i,j)th entry

FluxDiffUtils.hadamard_sum! — Method

hadamard_sum!(rhs, A_list::NTuple{N,T}, F::Fxn, u, Fargs...;
             skip_index=(i,j)->false) where {N,T,Fxn}

Mutating version of hadamard_sum, where rhs is storage for output.

FluxDiffUtils.hadamard_sum — Method

hadamard_sum(A_list::NTuple{N,T}, F::Fxn, u, Fargs...;
             skip_index=(i,j)->false) where {N,T,Fxn}

computes ∑i sum(Ai.*Fi,dims=2) where (Fi)jk = F(uj,uk)[i]

Inputs

A_list: tuple of operators (A1,...,Ad)

F: flux function which outputs a d-tuple of flux vectors

u: collection of solution values (or arrays) at which to evaluate F

Fargs: extra arguments to F(ui,getindex.(Fargs,i)..., uj,getindex.(Fargs,j)...)`

(optional) skip_index(i,j)==true skips computing fluxes for index (i,j)